refactor(links) abstracted list component
