+++ src/components/links/LinksList.module.css
@@ -1,0 +1,15 @@
+.links-container {
+  height: calc(100% - 8.35rem);
+  overflow-y: auto;
+  padding: 0 0.5rem;
+  scroll-snap-type: y mandatory;
+
+  /* Hide scrollbar for IE, Edge and Firefox */
+  -ms-overflow-style: none; /* IE and Edge */
+  scrollbar-width: none; /* Firefox */
+}
+
+.links-container::-webkit-scrollbar {
+  /* Hide scrollbar for Chrome, Safari and Opera */
+  display: none;
+}
